load npcs; send list of npc to player; change gm teleport to z coordinate

+++ src/Imgeneus.World/Game/Zone/MapConfig/NpcCoordinate.cs
@@ -14,6 +14,6 @@ namespace Imgeneus.World.Game.Zone.MapConfig
         public float Z { get; set; }
 
         [JsonPropertyName("angle")]
-        public ushort Angle { get; set; }
+        public float Angle { get; set; }
     }
 }
